BUG/MINOR: checks: Compute the right HTTP request length for HTTP health checks
A bug was introduced by the commit 104d66a ("MINOR: checks: Add a way to send
custom headers and payload during http chekcs"). When the method and the uri are
configured for an HTTP health check, but not the version, the request length is
not properly computed. One byte is missing (the space between the method and the
uri).

This patch should fix the issue #641. The patch was directly introduced on 2.1,
there is no upstream commit ID for this patch. It must be backported everywhere
the abobe commit is.

(cherry picked from commit f6f9ae40f60dd6de41809db19c28fab8afc9995e)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>