Revert "BUG/MEDIUM: connections: force connections cleanup on server changes"
As explained by Christopher on github issue #665:
In 2.2, srv->idle_conns and srv->safe_conns are thread-safe lists. But
not in 2.1. So the patch must be reverted or the lists must be changed
to use mt_list instead. The same must be done in 2.0, but the mt_list
does not exist on this version.

I choose to revert it as the original bug is truly revealed in v2.2
after commit 079cb9af22da6 ("MEDIUM: connections: Revamp the way idle
connections are killed")

this should resolve github issue #665

this reverts commit 7eab37b6819af685c647cf5a581e29fca2f3e079.
this reverts commit 3ad3306ec0bcb0cd4ca2b9ba134ed67663473ee8.

The patch was directly introduced on 2.1, there is no upstream commit ID for
this patch.

(cherry picked from commit a31179b920929e30f2ec970bf3a3dc26d9ae50cd)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
[Cf: reverted commit ids on 2.0 are e18f603012fe3b698a30fcedfb0684f48fe403aa and
     3b614335cf25a7081c25e0a789c932c5960f221c. If possible don't backport them
     to 1.9]