BUG/MINOR: proto-http: Fix detection of NTLM for the legacy HTTP version
The detection of NTLM was recently fixed. In mainline and the 2.1, only the HTX
part was concerned because the legacy HTTP part no longer exists. But we forgot
to update the legacy HTTP code (or we failed to do it properly) during the
backports to the 2.0. Following commits are concerned:

  * 9df188695 BUG/MEDIUM: http-ana: Handle NTLM messages correctly.
  * 49a1d28fc BUG/MINOR: http-ana: fix NTLM response parsing again
  * f1dccedcf BUG/MEDIUM: http_ana: make the detection of NTLM variants safer

In legacy HTTP code, the matching must not be performed using an exact word
matching function. Only prefix "Nego" or "NTLM" must be checked.

The patch was directly introduced on 2.0, there is no upstream commit ID for
this patch. It must be backported with the commits above.