BUG/MEDIUM: mux-h2: Don't add private connections in available connection list
When a stream is detached from a backend private connection, we must not insert
it in the available connection list. In addition, we must be sure to remove it
from this list. To ensure it is properly performed, this part has been slightly
refactored to clearly split processing of private connections from the others.

This patch should probably be backported to 2.2.

(cherry picked from commit c5579d18d1a5161beaeda57705809ef5434bb9dd)
Signed-off-by: Willy Tarreau <w@1wt.eu>